Add CI quality gates for lint and coverage
Split ci.yml into lint/test/coverage jobs and turn the five lint metrics
into real gates: --fail/--max-count support in the lint scripts, a
lint:gates aggregation script, eslint --max-warnings baseline, knip.json
entry declarations, and c8 check-coverage thresholds. To make the gates
hold, split automation-service.ts (998 lines) into automation-event-utils,
extract the shared list-query helper in security-store, and share the
Knowledge notice banner component.
