+++ package.json
@@ -38,11 +38,13 @@
     "test:bdd": "node --import tsx scripts/test-bdd.mjs --config tests/bdd/cucumber.mjs",
     "coverage": "node scripts/coverage.mjs",
     "lint:complexity": "eslint . --format stylish",
+    "lint:complexity:gate": "eslint . --format stylish --max-warnings 108",
     "lint:circular": "node scripts/lint-circular.mjs",
     "lint:dead-code": "node scripts/lint-dead-code.mjs",
     "lint:duplicate": "node scripts/lint-duplicate.mjs",
     "lint:file-size": "node scripts/lint-file-size.mjs",
     "lint:function-length": "node scripts/lint-function-length.mjs",
+    "lint:gates": "node scripts/lint-circular.mjs --fail && JSCPD_MIN_LINES=10 JSCPD_MIN_TOKENS=60 node scripts/lint-duplicate.mjs --fail && node scripts/lint-dead-code.mjs --fail --max-count 171 && node scripts/lint-file-size.mjs --fail && node scripts/lint-function-length.mjs --fail --max-count 45 && pnpm lint:complexity:gate",
     "e2e:smoke": "pnpm build && node scripts/e2e-smoke.mjs",
     "e2e:compose": "node scripts/e2e-compose.mjs",
     "dist:mac": "pnpm build && electron-builder --mac dmg zip --arm64",
